plan(06-arc-A1): clean clippy debt (40 pre-existing branch errors → 0)
The branch had accumulated 40 clippy errors from spine commits 87c2ce8
onwards (Stitch::Discard, Queue, ref-queue, Purity analyser groundwork,
ParFor walker arms, etc.) that no commit gated through `cargo clippy
-D warnings`.  A1's "make ci clean" acceptance gate caught the debt;
this commit pays it down so A1 closes cleanly.

Categories cleared:

- 9× doc list overindented (src/data.rs:379-391 — re-flowed
  continuation lines under the matching list-item indent)
- 6× cast u16→u32 (clippy autofix → `From::from`)
- 4× unnested or-patterns in src/scopes.rs Purity match arms
  (clippy autofix folded `Purity::Pure | Purity::Impure(A) |
  Purity::Impure(B)` into a nested-or)
- 4× cast u16→i32 (autofix)
- 2× clippy::too_many_arguments (allow on
  `run_parallel_discard` and `handle_operator` — both legitimately
  need the args; suppressing locally is preferable to redesigning
  function signatures already tested under the spine work)
- 2× clippy::similar_names (allow on `collect_callees` for
  callee/caller pair, and `check_fields_overlap` for a/b pair)
- 2× same-type cast (autofix)
- 1× format!()-appended-to-String → use `write!` (src/database/types.rs)
- 1× missing # Panics in src/introspect.rs::run (added)
- 1× clippy::assigning_clones → `clone_from`
- 1× many_single_char_names allow on `validate_layout_by_nr`
  (legitimate single-char usage in field-iteration code)
- 1× length-comparison-to-one in tests/parallel_rebase.rs →
  `is_empty()` negation
- 2× doc list without indentation in tests/threading{,_chars}.rs
  (rephrased "+" continuation as "and"/"plus")
- 1× sort→sort_unstable on u16 in test code

No semantic changes; all edits are lint-driven style fixes.

`cargo fmt --check` clean.  `cargo clippy --all-targets -- -D
warnings` clean.  `cargo build --no-default-features` clean.

`cargo nextest run --profile ci` reveals one remaining test failure
(`loft::html_wasm::moros_editor_html_smoke`) which is a pre-existing
P199 / A7 manifestation (E0499 double-borrow in
`OpCopyRecord(stores, n_build_chunk(stores, ...), ...)`) — the same
codegen bug A1's bench-11 native column hit and the same one ARC
step A7 closes.  Not an A1 regression.

+++ src/parser/mod.rs
@@ -1960,7 +1960,7 @@ impl Parser {
                 // and `output_call` projects `.0` natively (see the
                 // val_is_fn_ref handling in src/generation/calls.rs).
                 let d_nr_only = match value {
-                    Value::FnRef(d_nr, _, _) => Value::Int(d_nr as i32),
+                    Value::FnRef(d_nr, _, _) => Value::Int(d_nr),
                     other => other,
                 };
                 self.cl("OpSetInt4", &[ref_code.clone(), pos_v, d_nr_only])
@@ -2150,7 +2150,7 @@ impl Parser {
                 // `src/generation/calls.rs` projects `.0` from the
                 // `(u32, DbRef)` fn-ref tuple before the i32 cast.
                 let d_nr_only = match val_code {
-                    Value::FnRef(d_nr, _, _) => Value::Int(d_nr as i32),
+                    Value::FnRef(d_nr, _, _) => Value::Int(d_nr),
                     other => other,
                 };
                 self.cl("OpSetInt4", &[ref_code, pos_val, d_nr_only])
